New submodule IO for I/O operations
