+++ src/ParticlesMC.jl
@@ -25,6 +25,6 @@ export fold_back, System
 export SimpleGaussian, DoubleUniform, EnergyBias
 export sample_action!, log_proposal_density, reward, invert_action!, delta_log_target_density
 export perform_action!, perform_action_cached!
-export load_configuration, load_init_files
 
+include("IO/IO.jl")
 end
