+++ test/gerhard_energy_distribution.jl
@@ -1,6 +1,7 @@
 # TEST WITH GERHARD'S CONFIG
 using MonteCarlo
 using ParticlesMC
+using ParticlesMC.IO
 using StaticArrays
 using Distributions
 using Random
@@ -55,10 +56,10 @@ pools = [(
 
 algorithm_list = (
     (algorithm=Metropolis, pools=pools, seed=seed, parallel=false, sweepstep=N),
-    (algorithm=StoreCallbacks, callbacks=(callback_energy, callback_acceptance), scheduler=sampletimes),
-    (algorithm=StoreTrajectories, scheduler=sampletimes),
-    (algorithm=StoreLastFrames, scheduler=[steps]),
-    (algorithm=PrintTimeSteps, scheduler=build_schedule(steps, burn, steps ÷ 10)),
+    (algorithm=StoreCallbacks, callbacks=(callback_energy, callback_acceptance), scheduler=sampletimes, fmt=XYZ()),
+    (algorithm=StoreTrajectories, scheduler=sampletimes, fmt=XYZ()),
+    (algorithm=StoreLastFrames, scheduler=[steps], fmt=XYZ()),
+    (algorithm=PrintTimeSteps, scheduler=build_schedule(steps, burn, steps ÷ 10), fmt=XYZ()),
 )
 
 ## Empty List
